+++ gendarme/rules/Gendarme.Rules.Smells/CodeDuplicatedLocator.cs
@@ -31,6 +31,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.Globalization;
+using System.Linq;
 
 using Mono.Cecil;
 using Mono.Cecil.Cil;
@@ -40,18 +41,26 @@ using Gendarme.Framework.Rocks;
 
 namespace Gendarme.Rules.Smells {
 
+    public enum DetectionMode
+    {
+        Classic,
+        Modern
+    }
+
 	internal sealed class CodeDuplicatedLocator {
 
 		static ReadOnlyCollection<Pattern> Empty = new ReadOnlyCollection<Pattern> (new List<Pattern> ());
+        private readonly DetectionMode mode;
 
 		HashSet<string> methods = new HashSet<string> ();
 		HashSet<string> types = new HashSet<string> ();
 		Dictionary<MethodDefinition, IList<Pattern>> patternsCached = new Dictionary<MethodDefinition, IList<Pattern>> ();
 		IRule parent_rule;
 
-		internal CodeDuplicatedLocator (IRule rule) 
+		internal CodeDuplicatedLocator (IRule rule, DetectionMode kind) 
 		{
 			parent_rule = rule;
+            mode = kind;
 		}
 
 		internal ICollection<string> CheckedMethods {
@@ -84,7 +93,11 @@ namespace Gendarme.Rules.Smells {
 				Pattern duplicated = GetDuplicatedCode (current, target);
 				if (duplicated != null && duplicated.Count > 0) {
 					parent_rule.Runner.Report (current, duplicated[0], Severity.High, Confidence.Normal, 
-						String.Format (CultureInfo.InvariantCulture, "Duplicated code with {0}", target.GetFullName ()));
+						String.Format (CultureInfo.InvariantCulture, "Duplicated code with {0}{1}{2}{3}",
+                        mode == DetectionMode.Classic ? String.Empty : Environment.NewLine,
+                        mode == DetectionMode.Classic ? String.Empty : duplicated.ToString(),
+                        mode == DetectionMode.Classic ? String.Empty : Environment.NewLine,
+                        target.GetFullName()));
 				}
 			}
 		}
@@ -101,11 +114,11 @@ namespace Gendarme.Rules.Smells {
 		{
 			Log.WriteLine (this, "Found pattern in {0} and {1}", current, target);
 			Log.WriteLine (this, "\t Pattern");
-			for (int index = 0; index < found.Count; index++) 
+			for (int index = 0; index < found.Count; index++)
 				Log.WriteLine (this, "\t\t{0} - {1}",
-					found[index].OpCode.Code,
-					found[index].Operand != null? found[index].Operand : "No operator");
-		}
+                    found[index].OpCode.Code,
+                    found[index].Operand != null ? found[index].Operand : "No operator");
+        }
 
 		Pattern GetDuplicatedCode (MethodDefinition current, MethodDefinition target)
 		{
@@ -131,13 +144,13 @@ namespace Gendarme.Rules.Smells {
 
 			return null;
 		}
-		
+
 
 		IList<Pattern> GetPatterns (MethodDefinition method) 
 		{
 			IList<Pattern> patterns = Empty;
 			if (!patternsCached.TryGetValue (method, out patterns)) {
-				patterns = GeneratePatterns (method);
+				patterns = GeneratePatterns (method, this.mode);
 				patternsCached.Add (method, patterns);
 			}
 			return patterns;
@@ -146,7 +159,7 @@ namespace Gendarme.Rules.Smells {
 		//TODO: Still needs some testing in order to get the best size
 		//for every case:
 		//  The idea is get two overlapped statements in high level language
-		static IList<Pattern> GeneratePatterns (MethodDefinition method) 
+		static IList<Pattern> GeneratePatterns (MethodDefinition method, DetectionMode mode) 
 		{
 			Stack<Stack<Instruction>> result = new Stack<Stack<Instruction>> ();
 			Stack<Instruction> current = new Stack<Instruction> ();
@@ -176,9 +189,57 @@ namespace Gendarme.Rules.Smells {
 			if (result.Count == 0)
 				return Empty;
 
-			IList<Pattern> res = new List<Pattern> ();
-			foreach (Stack<Instruction> stack in result) 
-				res.Add (new Pattern (stack.ToArray ()));
+			List<Pattern> res =
+                result.Select(stack => stack.ToArray())
+                      .Where(stack =>
+                      {
+                          if (mode == DetectionMode.Classic)
+                              return true;
+
+                          // ignore "throw new ArgumentNullException" boilerplate
+                          if (stack.Any(inst =>
+                          {
+                              if (inst.OpCode == OpCodes.Throw &&
+                                  inst.Previous != null &&
+                                  inst.Previous.OpCode == OpCodes.Newobj &&
+                                  inst.Previous.Operand is MethodReference)
+                              {
+                                  var reference = (inst.Previous.Operand as MethodReference).FullName;
+                                  if (reference == "System.Void System.ArgumentNullException::.ctor(System.String)")
+                                      return true;
+                              }
+                              return false;
+                          })) return false;
+
+                          var dbg = method.DebugInformation;
+                          if (dbg == null)
+                              return true;
+
+                          var sp = dbg.GetSequencePoint(stack[0]);
+                          if (sp == null)
+                              return false; // patterns start at a SeqPnt
+
+                          var last = stack.Last();
+                          var next = last.Next;
+                          if (next != null && dbg.GetSequencePoint(next) == null)
+                              return false; // patterns end at the end of a SeqPnt
+
+                          if (stack.Any(inst =>
+                          {
+                              var ss = dbg.GetSequencePoint(inst);
+                              return ss != null && ss.IsHidden;
+                          }))
+                              return false; // patterns don't include hidden lines
+
+                          //// patterns are 2 or more SeqPnts in length (i.e. are not trivial)
+                          //var sp = stack.Where(i => dbg.GetSequencePoint(i)!= null).Count();
+                          //return (sp > 1);
+                          return true;
+                      })
+                      .Select(array => new Pattern(array, method))
+                      .ToList();
+
+//            res.ForEach(p => WriteToOutput(method, p));
 
 			return res;
 		}